chore: clean up enums for adjusting traits
